refactor(game): extract round-timer + REVEAL/terminal transition into RevealTransitionMixin (#1271)
Next increment of the state.py God-Object split (#1271). Pulls the
forward-flow machinery that carries a round from "all guesses in / timer
expired" through REVEAL to the terminal END phase out of GameState into a
new RevealTransitionMixin (game/state_reveal_transition.py):

- early-reveal gate: check_all_guesses_complete, _trigger_early_reveal,
  trigger_early_reveal_if_complete
- round timer: _timer_countdown, cancel_timer
- REVEAL transition helpers: _transition_to_reveal, _apply_reveal_lights
- intro-splash / deadline delegations: confirm_intro_splash, is_deadline_passed
- terminal transition: advance_to_end

Pure code-move, byte-for-byte API-identical. The round-end / phase SSOT
stays on GameState: _set_phase chokepoint (+ _VALID_PHASE_TRANSITIONS /
_UNIVERSAL_PHASE_TARGETS), _score_all_players, _end_round_unlocked and
_schedule_reveal_advance. Mixin added alphabetically to the GameState base
list + docstring paragraph.

state.py: 1078 -> 898 lines (-180).

Gates: 885 passed / 2 skipped, ruff check + format clean, mypy Success.
